+++ packages/stream_chat/lib/src/core/models/channel_model.dart
@@ -14,7 +14,7 @@ class ChannelModel {
     String? id,
     String? type,
     String? cid,
-    this.ownCapabilities,
+    List<String>? ownCapabilities,
     ChannelConfig? config,
     this.createdBy,
     this.frozen = false,
@@ -40,6 +40,7 @@ class ChannelModel {
         config = config ?? ChannelConfig(),
         createdAt = createdAt ?? DateTime.now(),
         updatedAt = updatedAt ?? DateTime.now(),
+        ownCapabilities = ownCapabilities?.map(ChannelCapability.new).toList(),
 
         // For backwards compatibility, set 'disabled', 'hidden'
         // and 'truncated_at' in [extraData].
@@ -67,9 +68,9 @@ class ChannelModel {
   @JsonKey(includeToJson: false)
   final String cid;
 
-  /// List of user permissions on this channel
+  /// List of various capabilities that a user can have in a channel.
   @JsonKey(includeToJson: false)
-  final List<String>? ownCapabilities;
+  final List<ChannelCapability>? ownCapabilities;
 
   /// The channel configuration data
   @JsonKey(includeToJson: false)
@@ -237,4 +238,119 @@ class ChannelModel {
       truncatedAt: other.truncatedAt,
     );
   }
+}
+
+/// {@template channelCapability}
+/// Represents various capabilities that a user can have in a channel.
+/// {@endtemplate}
+extension type const ChannelCapability(String capability) implements String {
+  /// Ability to send a message.
+  static const sendMessage = ChannelCapability('send-message');
+
+  /// Ability to reply to a message.
+  static const sendReply = ChannelCapability('send-reply');
+
+  /// Ability to send a message with restricted visibility.
+  static const sendRestrictedVisibilityMessage = ChannelCapability(
+    'send-restricted-visibility-message',
+  );
+
+  /// Ability to send reactions.
+  static const sendReaction = ChannelCapability('send-reaction');
+
+  /// Ability to attach links to messages.
+  static const sendLinks = ChannelCapability('send-links');
+
+  /// Ability to attach files to messages.
+  static const createAttachment = ChannelCapability('create-attachment');
+
+  /// Ability to freeze or unfreeze channel.
+  static const freezeChannel = ChannelCapability('freeze-channel');
+
+  /// Ability to enable or disable slow mode.
+  static const setChannelCooldown = ChannelCapability('set-channel-cooldown');
+
+  /// Ability to leave channel (remove own membership).
+  static const leaveChannel = ChannelCapability('leave-channel');
+
+  /// Ability to join channel (add own membership).
+  static const joinChannel = ChannelCapability('join-channel');
+
+  /// Ability to pin a message.
+  static const pinMessage = ChannelCapability('pin-message');
+
+  /// Ability to delete any message from the channel.
+  static const deleteAnyMessage = ChannelCapability('delete-any-message');
+
+  /// Ability to delete own messages from the channel.
+  static const deleteOwnMessage = ChannelCapability('delete-own-message');
+
+  /// Ability to update any message in the channel.
+  static const updateAnyMessage = ChannelCapability('update-any-message');
+
+  /// Ability to update own messages in the channel.
+  static const updateOwnMessage = ChannelCapability('update-own-message');
+
+  /// Ability to use message search.
+  static const searchMessages = ChannelCapability('search-messages');
+
+  /// Ability to send typing events.
+  static const sendTypingEvents = ChannelCapability('send-typing-events');
+
+  /// Ability to upload message attachments.
+  static const uploadFile = ChannelCapability('upload-file');
+
+  /// Ability to delete channel.
+  static const deleteChannel = ChannelCapability('delete-channel');
+
+  /// Ability to update channel data.
+  static const updateChannel = ChannelCapability('update-channel');
+
+  /// Ability to update channel members.
+  static const updateChannelMembers = ChannelCapability(
+    'update-channel-members',
+  );
+
+  /// Ability to update thread data.
+  static const updateThread = ChannelCapability('update-thread');
+
+  /// Ability to quote a message.
+  static const quoteMessage = ChannelCapability('quote-message');
+
+  /// Ability to ban channel members.
+  static const banChannelMembers = ChannelCapability('ban-channel-members');
+
+  /// Ability to flag a message.
+  static const flagMessage = ChannelCapability('flag-message');
+
+  /// Ability to mute a channel.
+  static const muteChannel = ChannelCapability('mute-channel');
+
+  /// Ability to send custom events.
+  static const sendCustomEvents = ChannelCapability('send-custom-events');
+
+  /// Ability to receive read events.
+  static const readEvents = ChannelCapability('read-events');
+
+  /// Ability to receive connect events.
+  static const connectEvents = ChannelCapability('connect-events');
+
+  /// Ability to send and receive typing events.
+  static const typingEvents = ChannelCapability('typing-events');
+
+  /// Indicates that channel slow mode is active.
+  static const slowMode = ChannelCapability('slow-mode');
+
+  /// Indicates that the user is allowed to post messages as usual even if the
+  /// channel is in slow mode.
+  static const skipSlowMode = ChannelCapability('skip-slow-mode');
+
+  /// Ability to update a poll.
+  static const sendPoll = ChannelCapability('send-poll');
+
+  /// Ability to update a poll.
+  static const castPollVote = ChannelCapability('cast-poll-vote');
+
+  /// Ability to query poll votes.
+  static const queryPollVotes = ChannelCapability('query-poll-votes');
 }